requirements.txt
aiosqlite>=0.19.0,<1.0.0
websockets>=12.0,<13.0
pytest
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
pytest-asyncio>=0.23.0
